refactor(validated input): extract ValidatedInput from Validator to an own file
